@@ -1,8 +1,9 @@
 package com.example.a14512.audiovediodemo.camera2;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.example.a14512.audiovediodemo.R;
 
