@@ -3,11 +3,12 @@ package com.example.a14512.audiovediodemo;
 import android.content.Intent;
 import android.os.Build;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
 import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 
 import com.example.a14512.audiovediodemo.camera2.Camera2Demo1Activity;
 import com.example.a14512.audiovediodemo.camera2.Camera2DemoActivity;
